requirements.txt
Pillow>=10.0.0
matplotlib>=3.0.0
